--- checkout_page.dart
@@ -1,28 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CheckoutPage extends StatefulWidget {
-  @override
-  _CheckoutPageState createState() => _CheckoutPageState();
-}
-
-class _CheckoutPageState extends State<CheckoutPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        centerTitle: true,
-        title: Text(
-          'CheckOut',
-          style: TextStyle(
-            color: Colors.black,
-          ),
-        ),
-        backgroundColor: Colors.white,
-        brightness: Brightness.light,
-        elevation: 0,
-        actionsIconTheme: IconThemeData(color: Colors.black),
-        iconTheme: IconThemeData(color: Colors.black),
-      ),
-    );
-  }
-}
