+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:fiska/pages/homePage.dart';
+import 'package:fiska/splash_screen.dart';
 import 'package:flutter/material.dart';
 
 void main() => runApp(App());
@@ -8,9 +8,11 @@ class App extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
+      home: SplashScreen(),
       theme: ThemeData(
-          fontFamily: "poppins", scaffoldBackgroundColor: Colors.white),
-      home: HomePage(),
+        fontFamily: "poppins",
+        scaffoldBackgroundColor: Colors.white,
+      ),
     );
   }
 }
