+++ lib/pages/homePage.dart
@@ -1,13 +1,16 @@
+import 'package:badges/badges.dart';
 import 'package:carousel_slider/carousel_slider.dart';
 import 'package:curved_navigation_bar/curved_navigation_bar.dart';
 import 'package:eva_icons_flutter/eva_icons_flutter.dart';
-import 'package:fiska/models/product.dart';
-import 'package:fiska/pages/profilePage.dart';
-import 'package:fiska/widgets/categoryItem.dart';
-import 'package:fiska/pages/productPage.dart';
-import 'package:fiska/pages/categoryPage.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+
+import 'package:fiska/models/product.dart';
+import 'package:fiska/pages/categoryPage.dart';
+import 'package:fiska/pages/productPage.dart';
+import 'package:fiska/pages/profilePage.dart';
+import 'package:fiska/widgets/categoryItem.dart';
+
 import 'SearchPage.dart';
 import 'cartPage.dart';
 
@@ -115,9 +118,9 @@ class _HomePageState extends State<HomePage> {
         appBar: AppBar(
           centerTitle: true,
           title: Text(
-            "Fiska",
+            "FISKA",
             style: TextStyle(
-                color: Colors.deepOrange,
+                color: Colors.orange[800],
                 fontSize: 26.0,
                 fontWeight: FontWeight.bold,
                 letterSpacing: 5.0),
@@ -125,8 +128,8 @@ class _HomePageState extends State<HomePage> {
           backgroundColor: Colors.white,
           brightness: Brightness.light,
           elevation: 0,
-          actionsIconTheme: IconThemeData(color: Colors.deepOrange),
-          iconTheme: IconThemeData(color: Colors.deepOrange),
+          actionsIconTheme: IconThemeData(color: Colors.orange),
+          iconTheme: IconThemeData(color: Colors.orange),
           actions: <Widget>[
             IconButton(
               onPressed: () {
@@ -134,10 +137,27 @@ class _HomePageState extends State<HomePage> {
               },
               icon: Icon(EvaIcons.search),
             ),
-            IconButton(
-              onPressed: () {},
-              icon: Icon(EvaIcons.shoppingBagOutline),
-            ),
+            Badge(
+              padding: EdgeInsets.all(2),
+              shape: BadgeShape.circle,
+              borderRadius: 3,
+              badgeContent: Text(
+                '10',
+                style:
+                    TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
+              ),
+              position: BadgePosition.topEnd(top: 5, end: 4),
+              animationDuration: Duration(milliseconds: 300),
+              animationType: BadgeAnimationType.slide,
+              child: IconButton(
+                onPressed: () {},
+                icon: Icon(
+                  Icons.shopping_cart,
+                  size: 22,
+                  color: Colors.orange,
+                ),
+              ),
+            )
           ],
         ),
         drawerEdgeDragWidth: 0,
@@ -175,21 +195,25 @@ class _HomePageState extends State<HomePage> {
               ListTile(
                 title: Text("Account"),
                 leading: Icon(EvaIcons.personOutline),
+                onTap: () {},
               ),
               SizedBox(height: 10),
               ListTile(
                 title: Text("Categories"),
                 leading: Icon(Icons.list),
+                onTap: () {},
               ),
               SizedBox(height: 10),
               ListTile(
                 title: Text("Sevices"),
                 leading: Icon(Icons.live_help),
+                onTap: () {},
               ),
               SizedBox(height: 10),
               ListTile(
                 title: Text("Help"),
                 leading: Icon(EvaIcons.bulbOutline),
+                onTap: () {},
               ),
               SizedBox(height: 10),
               ListTile(
@@ -319,7 +343,7 @@ class _HomePageState extends State<HomePage> {
 
                   CarouselSlider(
                     options: CarouselOptions(
-                      aspectRatio: 16 / 9,
+                      aspectRatio: 16 / 7,
                       autoPlay: true,
                     ),
                     items: bannerAdSlider.map((i) {
@@ -329,7 +353,7 @@ class _HomePageState extends State<HomePage> {
                             width: MediaQuery.of(context).size.width,
                             margin: EdgeInsets.symmetric(horizontal: 2.0),
                             child: ClipRRect(
-                              borderRadius: BorderRadius.circular(16),
+                              borderRadius: BorderRadius.circular(30),
                               child: Image(
                                 image: AssetImage(i),
                                 fit: BoxFit.cover,
@@ -356,99 +380,109 @@ class _HomePageState extends State<HomePage> {
                           TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
                     ),
                   ),
-
+                  Divider(
+                    height: 3.0,
+                  ),
                   SizedBox(
-                    height: 20,
+                    height: 8,
                   ),
 
                   GridView.count(
-                    padding: EdgeInsets.all(4.0),
+                    padding: EdgeInsets.all(5.0),
                     physics: ClampingScrollPhysics(),
                     crossAxisCount: 2,
                     shrinkWrap: true,
-                    childAspectRatio: 1 / 1.30,
+                    childAspectRatio: 1 / 1.0,
                     children: products.map((product) {
-                      return Stack(
-                        children: <Widget>[
-                          Container(
-                            child: Column(
-                              children: <Widget>[
-                                /**Positioned(
-                                  top: 6.0,
-                                  right: 6.0,
-                                  child: Card(
-                                    color: Colors.red,
-                                    elevation: 2.0,
-                                    shape: RoundedRectangleBorder(
-                                        borderRadius:
-                                            BorderRadius.circular(4.0)),
-                                    child: Padding(
-                                      padding: EdgeInsets.all(9.0),
-                                      child: Row(
-                                        children: <Widget>[
-                                          Icon(
-                                            Icons.star,
-                                            color: Product.ratingBG,
-                                            size: 20,
-                                          ),
-                                          SizedBox(
-                                            width: 5.0,
-                                          ),
-                                          Text(
-                                            "${product.rating} ",
-                                            style: TextStyle(
-                                              fontSize: 20,
-                                              color: Colors.white,
-                                            ),
-                                          ),
-                                        ],
+                      return Card(
+                        elevation: 1.5,
+                        child: Stack(
+                          children: <Widget>[
+                            Container(
+                              child: Column(
+                                children: <Widget>[
+                                  Expanded(
+                                    child: Hero(
+                                      tag: product.image,
+                                      child: AspectRatio(
+                                        aspectRatio: 1 / 1,
+                                        child: Image(
+                                          image: AssetImage(product.image),
+                                        ),
                                       ),
                                     ),
                                   ),
-                                ),
-                                */
-                                Expanded(
-                                  child: Hero(
-                                    tag: product.image,
-                                    child: AspectRatio(
-                                      aspectRatio: 1 / 1,
-                                      child: Image(
-                                        image: AssetImage(product.image),
-                                      ),
+                                  Divider(
+                                    height: 3.0,
+                                  ),
+                                  Text(
+                                    product.productName,
+                                  ),
+                                  Text(
+                                    "${product.price}\$",
+                                    style: TextStyle(
+                                      fontSize: 18,
+                                      fontWeight: FontWeight.w700,
+                                      color: Colors.amber[900],
                                     ),
                                   ),
-                                ),
-                                Text(
-                                  product.productName,
-                                ),
-                                Text(
-                                  "${product.price}\$",
-                                  style: TextStyle(
-                                    fontSize: 18,
-                                    fontWeight: FontWeight.w700,
-                                    color: Colors.amber,
+                                ],
+                              ),
+                            ),
+                            Material(
+                              color: Colors.transparent,
+                              child: InkWell(
+                                onTap: () {
+                                  Navigator.push(
+                                      context,
+                                      MaterialPageRoute(
+                                        builder: (context) => ProductPage(
+                                          product: product,
+                                        ),
+                                      ));
+                                },
+                              ),
+                            ),
+                            Positioned(
+                              top: -2.0,
+                              right: -3.0,
+                              child: Card(
+                                color: Colors.amber[200],
+                                shape: RoundedRectangleBorder(
+                                    borderRadius: BorderRadius.circular(4.0)),
+                                child: Padding(
+                                  padding: EdgeInsets.all(1.0),
+                                  child: Row(
+                                    children: <Widget>[
+                                      Text(
+                                        '-70%',
+                                        style: TextStyle(
+                                            color: Colors.black,
+                                            fontSize: 16.0,
+                                            fontWeight: FontWeight.bold),
+                                      )
+                                    ],
                                   ),
                                 ),
-                              ],
+                              ),
                             ),
-                          ),
-                          Material(
-                            color: Colors.transparent,
-                            child: InkWell(
-                              onTap: () {
-                                Navigator.push(
-                                    context,
-                                    MaterialPageRoute(
-                                      builder: (context) => ProductPage(
-                                        product: product,
-                                      ),
-                                    ));
-                              },
-                            ),
-                          )
-                        ],
+                          ],
+                        ),
                       );
                     }).toList(),
+                  ),
+
+                  SizedBox(
+                    height: 8,
+                  ),
+
+                  Padding(
+                    padding: const EdgeInsets.all(8.0),
+                    child: Text(
+                      "Top Products",
+                      style:
+                          TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
+                    ),
                   ),
                 ],
               ),
@@ -472,10 +506,23 @@ class _HomePageState extends State<HomePage> {
                 size: 24,
                 color: Colors.white,
               ),
-              Icon(
-                Icons.shopping_cart,
-                size: 24,
-                color: Colors.white,
+              Badge(
+                padding: EdgeInsets.all(3),
+                shape: BadgeShape.circle,
+                borderRadius: 3,
+                badgeContent: Text(
+                  '10',
+                  style: TextStyle(
+                      color: Colors.white, fontWeight: FontWeight.bold),
+                ),
+                position: BadgePosition.topEnd(top: -10, end: -15),
+                animationDuration: Duration(milliseconds: 300),
+                animationType: BadgeAnimationType.slide,
+                child: Icon(
+                  Icons.shopping_cart,
+                  size: 24,
+                  color: Colors.white,
+                ),
               ),
               Icon(
                 Icons.person,
