+++ lib/pages/productPage.dart
@@ -1,7 +1,14 @@
 import 'package:eva_icons_flutter/eva_icons_flutter.dart';
+import 'package:fiska/controllers/cartcontroller.dart';
 import 'package:fiska/models/product.dart';
+import 'package:fiska/pages/cartPage.dart';
+import 'package:fiska/pages/checkout_page.dart';
+import 'package:flushbar/flushbar.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_rating_bar/flutter_rating_bar.dart';
+import 'package:get/get.dart';
+
+import 'SearchPage.dart';
 
 class ProductPage extends StatelessWidget {
   final Product product;
@@ -10,39 +17,51 @@ class ProductPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    CartModel _cartModel = Get.put(CartModel());
     return Scaffold(
       appBar: AppBar(
         centerTitle: true,
         title: Text(
           'Details',
           style: TextStyle(
-            color: Colors.white,
+            color: Colors.black,
           ),
         ),
-        backgroundColor: Colors.black87,
+        backgroundColor: Colors.white,
         brightness: Brightness.light,
         elevation: 0,
-        actionsIconTheme: IconThemeData(color: Colors.white),
-        iconTheme: IconThemeData(color: Colors.white),
+        actionsIconTheme: IconThemeData(color: Colors.black),
+        iconTheme: IconThemeData(color: Colors.orange),
         actions: <Widget>[
           Padding(
-              padding: EdgeInsets.only(right: 20.0),
-              child: GestureDetector(
-                onTap: () {},
-                child: Icon(
-                  Icons.search,
-                  size: 26.0,
-                ),
-              )
+            padding: EdgeInsets.only(right: 20.0),
+            child: GestureDetector(
+              onTap: () {
+                showSearch(context: context, delegate: ProductSearchDelegate());
+              },
+              child: Icon(
+                Icons.search,
+                size: 26.0,
+                color: Colors.orange,
+              ),
+            ),
           ),
           Padding(
-              padding: EdgeInsets.only(right: 20.0),
-              child: GestureDetector(
-                onTap: () {},
-                child: Icon(
-                    Icons.shopping_cart
-                ),
-              )
+            padding: EdgeInsets.only(right: 20.0),
+            child: GestureDetector(
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => CartPage(),
+                  ),
+                );
+              },
+              child: Icon(
+                Icons.shopping_cart,
+                color: Colors.orange,
+              ),
+            ),
           ),
         ],
       ),
@@ -66,10 +85,9 @@ class ProductPage extends StatelessWidget {
                 ),
                 child: ListTile(
                   title: Text('Seller'),
-                  onTap: (){},
+                  onTap: () {},
                 ),
               ),
-
               Padding(
                 padding: const EdgeInsets.symmetric(
                   horizontal: 10,
@@ -88,7 +106,7 @@ class ProductPage extends StatelessWidget {
                   vertical: 5,
                 ),
                 child: Text(
-                  "${product.price}\$",
+                  "\$${product.price}",
                   style: TextStyle(
                     fontSize: 50,
                     fontWeight: FontWeight.w700,
@@ -108,7 +126,6 @@ class ProductPage extends StatelessWidget {
                   ),
                 ),
               ),
-
               Padding(
                 padding: const EdgeInsets.symmetric(
                   horizontal: 20,
@@ -146,31 +163,30 @@ class ProductPage extends StatelessWidget {
                 ),
               ),
               Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 20,
-                  vertical: 0,
-                ),
-                child: RatingBar(
-                  initialRating: 3,
-                  minRating: 1,
-                  direction: Axis.horizontal,
-                  allowHalfRating: true,
-                  itemCount: 5,
-                  itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                  itemBuilder: (context, _) => Icon(
-                    Icons.star,
-                    color: Colors.amber,
+                  padding: const EdgeInsets.symmetric(
+                    horizontal: 20,
+                    vertical: 0,
                   ),
-                  onRatingUpdate: (rating) {
-                    print(rating);
-                  },
-                )
-              ),
+                  child: RatingBar(
+                    initialRating: 3,
+                    minRating: 1,
+                    direction: Axis.horizontal,
+                    allowHalfRating: true,
+                    itemCount: 5,
+                    itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
+                    itemBuilder: (context, _) => Icon(
+                      Icons.star,
+                      color: Colors.amber,
+                    ),
+                    onRatingUpdate: (rating) {
+                      print(rating);
+                    },
+                  )),
             ],
           ),
         ),
       ),
-      bottomNavigationBar:  Container(
+      bottomNavigationBar: Container(
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: <Widget>[
@@ -178,7 +194,9 @@ class ProductPage extends StatelessWidget {
               child: OutlineButton.icon(
                 icon: Icon(Icons.shopping_cart),
                 label: Text("Cart"),
-                onPressed: () {},
+                onPressed: () {
+                  _cartModel.inCart(product, context);
+                },
               ),
             ),
             SizedBox(
@@ -190,7 +208,14 @@ class ProductPage extends StatelessWidget {
                 color: Colors.amber,
                 icon: Icon(EvaIcons.creditCard),
                 label: Text("Buy Now"),
-                onPressed: () {},
+                onPressed: () {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (_) => CheckoutPage(),
+                    ),
+                  );
+                },
               ),
             ),
           ],
@@ -198,4 +223,38 @@ class ProductPage extends StatelessWidget {
       ),
     );
   }
+}
+
+void showSimpleFlushbar(BuildContext context, String message) {
+  Flushbar(
+    messageText: Text(
+      "$message",
+      style: TextStyle(
+        fontSize: 20.0,
+        fontWeight: FontWeight.w600,
+        color: Colors.black,
+      ),
+    ),
+    duration: Duration(seconds: 3),
+    mainButton: RaisedButton(
+      child: Text(
+        'Go to Cart',
+        style: TextStyle(
+          color: Colors.white,
+          fontSize: 18,
+          fontWeight: FontWeight.w400,
+        ),
+      ),
+      color: Colors.black12,
+      onPressed: () => Navigator.push(
+        context,
+        MaterialPageRoute(
+          builder: (context) => CartPage(),
+        ),
+      ),
+    ),
+    backgroundColor: Colors.orange,
+    padding: EdgeInsets.fromLTRB(20.0, 30.0, 20.0, 30.0),
+    borderRadius: 10.0,
+  )..show(context);
 }
