--- productPage.dart
@@ -1,252 +1,0 @@
-import 'package:eva_icons_flutter/eva_icons_flutter.dart';
-import 'package:fiska/controllers/cartcontroller.dart';
-import 'package:fiska/models/product.dart';
-import 'package:fiska/pages/cartPage.dart';
-import 'package:fiska/pages/checkout_page.dart';
-import 'package:flushbar/flushbar.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_rating_bar/flutter_rating_bar.dart';
-import 'package:get/get.dart';
-
-class ProductPage extends StatelessWidget {
-  final Product product;
-
-  ProductPage({this.product});
-
-  @override
-  Widget build(BuildContext context) {
-    CartModel _cartModel = Get.put(CartModel());
-    return Scaffold(
-      appBar: AppBar(
-        centerTitle: true,
-        title: Text(
-          'Details',
-          style: TextStyle(
-            color: Colors.black,
-          ),
-        ),
-        backgroundColor: Colors.white,
-        brightness: Brightness.light,
-        elevation: 0,
-        actionsIconTheme: IconThemeData(color: Colors.black),
-        iconTheme: IconThemeData(color: Colors.black),
-        actions: <Widget>[
-          Padding(
-            padding: EdgeInsets.only(right: 20.0),
-            child: GestureDetector(
-              onTap: () {},
-              child: Icon(
-                Icons.search,
-                size: 26.0,
-              ),
-            ),
-          ),
-          Padding(
-            padding: EdgeInsets.only(right: 20.0),
-            child: GestureDetector(
-              onTap: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => CartPage(),
-                  ),
-                );
-              },
-              child: Icon(Icons.shopping_cart),
-            ),
-          ),
-        ],
-      ),
-      body: Container(
-        child: SingleChildScrollView(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              Hero(
-                tag: product.image,
-                child: AspectRatio(
-                  aspectRatio: 1 / 1,
-                  child: Image(
-                    image: AssetImage(product.image),
-                  ),
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.only(
-                  top: 8,
-                ),
-                child: ListTile(
-                  title: Text('Seller'),
-                  onTap: () {},
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 10,
-                  vertical: 1,
-                ),
-                child: Text(
-                  product.productName,
-                  style: TextStyle(
-                    fontSize: 24,
-                  ),
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 10,
-                  vertical: 5,
-                ),
-                child: Text(
-                  "\$${product.price}",
-                  style: TextStyle(
-                    fontSize: 50,
-                    fontWeight: FontWeight.w700,
-                    color: Colors.amber,
-                  ),
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 10,
-                  vertical: 5,
-                ),
-                child: Text(
-                  'Coupons',
-                  style: TextStyle(
-                    fontSize: 24,
-                  ),
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 20,
-                  vertical: 0,
-                ),
-                child: Text(
-                  product.description,
-                  style: TextStyle(
-                    fontSize: 18,
-                  ),
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 20,
-                  vertical: 10,
-                ),
-                child: Row(
-                  children: [
-                    Text('Colors'),
-                    //The code for the variants should be written here
-                  ],
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.symmetric(
-                  horizontal: 20,
-                  vertical: 10,
-                ),
-                child: Text(
-                  'Reviews',
-                  style: TextStyle(
-                    fontSize: 18,
-                  ),
-                ),
-              ),
-              Padding(
-                  padding: const EdgeInsets.symmetric(
-                    horizontal: 20,
-                    vertical: 0,
-                  ),
-                  child: RatingBar(
-                    initialRating: 3,
-                    minRating: 1,
-                    direction: Axis.horizontal,
-                    allowHalfRating: true,
-                    itemCount: 5,
-                    itemPadding: EdgeInsets.symmetric(horizontal: 4.0),
-                    itemBuilder: (context, _) => Icon(
-                      Icons.star,
-                      color: Colors.amber,
-                    ),
-                    onRatingUpdate: (rating) {
-                      print(rating);
-                    },
-                  )),
-            ],
-          ),
-        ),
-      ),
-      bottomNavigationBar: Container(
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: <Widget>[
-            Expanded(
-              child: OutlineButton.icon(
-                icon: Icon(Icons.shopping_cart),
-                label: Text("Cart"),
-                onPressed: () {
-                  _cartModel.inCart(product, context);
-                },
-              ),
-            ),
-            SizedBox(
-              width: 10,
-            ),
-            Expanded(
-              child: RaisedButton.icon(
-                textColor: Colors.black,
-                color: Colors.amber,
-                icon: Icon(EvaIcons.creditCard),
-                label: Text("Buy Now"),
-                onPressed: () {
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(
-                      builder: (_) => CheckoutPage(),
-                    ),
-                  );
-                },
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-void showSimpleFlushbar(BuildContext context, String message) {
-  Flushbar(
-    messageText: Text(
-      "$message",
-      style: TextStyle(
-        fontSize: 20.0,
-        fontWeight: FontWeight.w600,
-        color: Colors.black,
-      ),
-    ),
-    duration: Duration(seconds: 3),
-    mainButton: RaisedButton(
-      child: Text(
-        'Go to Cart',
-        style: TextStyle(
-          color: Colors.white,
-          fontSize: 18,
-          fontWeight: FontWeight.w400,
-        ),
-      ),
-      color: Colors.black12,
-      onPressed: () => Navigator.push(
-        context,
-        MaterialPageRoute(
-          builder: (context) => CartPage(),
-        ),
-      ),
-    ),
-    backgroundColor: Colors.orange,
-    padding: EdgeInsets.fromLTRB(20.0, 30.0, 20.0, 30.0),
-    borderRadius: 10.0,
-  )..show(context);
-}
