+++ pubspec.yaml
@@ -24,13 +24,18 @@ dependencies:
   flutter:
     sdk: flutter
 
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.3
   curved_navigation_bar: ^0.3.4
   eva_icons_flutter: ^2.0.0
   carousel_slider: ^2.1.0
+  flutter_rating_bar: ^3.0.1+1
+  flushbar: ^1.10.4
+  get: ^3.12.1
+  convex_bottom_bar: ^2.5.1+1
+  animated_splash_screen: ^1.0.1+2
+  badges: ^1.1.3
 
 dev_dependencies:
   flutter_test:
@@ -38,7 +43,6 @@ dev_dependencies:
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-
 # The following section is specific to Flutter.
 flutter:
 
@@ -50,14 +54,10 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/
-  
-
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
-
   # For details regarding adding assets from package dependencies, see
   # https://flutter.dev/assets-and-images/#from-packages
-
   # To add custom fonts to your application, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
   # "family" key with the font family name, and a "fonts" key with a
@@ -77,3 +77,4 @@ flutter:
   #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
+
